requirements.txt
django==5.0.1 
djangorestframework==3.14.0 
django-cors-headers==4.3.1
djangorestframework-simplejwt == 5.5.0
django-filter == 25.1
python-dotenv==1.1.0
psycopg2-binary ==2.9.10
numpy ==2.2.6
drf-yasg==1.21.10
python-decouple==3.8
gunicorn==21.2.0
sqlparse==0.5.3
PyYAML==6.0.2
black == 25.1.0 
#uwsgi==2.0.30

# Para tarefas assíncronas
redis==6.4.0
celery==5.5.3
django-celery-beat==2.8.1



django-phonenumber-field[phonenumbers]
django-simple-history == 3.10.1

# No requirements.txt deve ter:
whitenoise==6.6.0

#django-debug-toolbar==5.2.0   Tool bar de performance

# Para validação de dados
django-extensions == 3.2.3
Pillow == 11.2.1

# Para exportação de dados
xlsxwriter == 3.2.0

requests == 2.31.0